8 6

+++ pointers_arrays_strings/8-print_array.c
@@ -13,7 +13,7 @@ void print_array(int *a, int n)
 
 	for (x = 0; x < n; x++)
 	{
-		printf("%d", a[x]);
+		printf("%ld", (long)a[x]);
 		if (x < - 1)
 		{
 			printf(", ");
